+++ example/example.dart
@@ -35,7 +35,7 @@ class _ExampleHomePageState extends State<ExampleHomePage>
 
   @override
   Widget build(BuildContext context) {
-    CardController controller;  //Use this to trigger swap.
+    CardController controller; //Use this to trigger swap.
 
     return new Scaffold(
       body: new Center(
@@ -45,6 +45,7 @@ class _ExampleHomePageState extends State<ExampleHomePage>
                   orientation: AmassOrientation.BOTTOM,
                   totalNum: 6,
                   stackNum: 3,
+                  swipeEdge: 4.0,
                   maxWidth: MediaQuery.of(context).size.width * 0.9,
                   maxHeight: MediaQuery.of(context).size.width * 0.9,
                   minWidth: MediaQuery.of(context).size.width * 0.8,
@@ -53,10 +54,17 @@ class _ExampleHomePageState extends State<ExampleHomePage>
                         child: Image.asset('${welcomeImages[index]}'),
                       ),
                   cardController: controller = CardController(),
-                  swipeUpdateCallback: (DragUpdateDetails details) {
-                    /// Get swiping card's position
+                  swipeUpdateCallback:
+                      (DragUpdateDetails details, Alignment align) {
+                    /// Get swiping card's alignment
+                    if (align.x < 0) {
+                      //Card is LEFT swiping
+                    } else if (align.x > 0) {
+                      //Card is RIGHT swiping
+                    }
                   },
-                  swipeCompleteCallback: (CardSwipeOrientation orientation, int index) {
+                  swipeCompleteCallback:
+                      (CardSwipeOrientation orientation, int index) {
                     /// Get orientation & index of swiped card!
                   }))),
     );